code for estimating protest intensity
code for estimating protest intensity, at the monthly level, using the
method described in our blog
(http://gdeltblog.wordpress.com/2013/12/30/more-on-normalizing-gdelt-dat
a/#comments).  also constructs the protest measure using the ratios as
described in Brandt's post
(http://gdeltblog.wordpress.com/2013/09/28/normalizing-gdelt-protest-dat
a/).  this is at the monthly and yearly levels.